+++ src/Products/PloneMeeting/content/person.py
@@ -2,6 +2,7 @@
 
 from collective.contact.core.content.person import IPerson
 from collective.contact.core.content.person import Person
+from Products.PloneMeeting.events import _invalidateAttendeesRelatedCache
 from plone.autoform import directives as form
 from plone.dexterity.schema import DexteritySchemaPolicy
 from plone.supermodel import model
@@ -44,6 +45,10 @@ class PMPerson(Person):
         for held_position in held_positions:
             if not position_type or held_position.position_type == position_type:
                 return held_position
+
+    def _invalidateCachedMethods(self):
+        '''Clean cache for vocabularies using held_positions.'''
+        _invalidateAttendeesRelatedCache()
 
 
 class PMPersonSchemaPolicy(DexteritySchemaPolicy):